+++ shared/src/commonMain/kotlin/com/services/network/api/NewsApi.kt
@@ -32,34 +32,9 @@ import io.ktor.serialization.kotlinx.json.json
 import kotlinx.serialization.json.Json
 
 internal class NewsApi(
-    private val baseAddress: String
+    private val baseAddress: String,
+    private val client: HttpClient
 ): BaseApi() {
-    private val client = HttpClient {
-        expectSuccess = true
-        install(ContentNegotiation) {
-            json(
-                Json {
-                    prettyPrint = true
-                    ignoreUnknownKeys = true
-                }
-            )
-        }
-        install(Logging) {
-            logger = Logger.SIMPLE
-            level = LogLevel.ALL
-        }
-        install(HttpTimeout)
-        install(Auth) {
-            basic {
-                credentials {
-                    BasicAuthCredentials(
-                        username = basicLogin,
-                        password = basicPassword
-                    )
-                }
-            }
-        }
-    }
 
     suspend fun getNewsByCategory(categoryName: String): ActionResult<NewsResponseListDTO, NetworkError> {
         return try {
